apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
redis==5.2.1
httpx==0.28.1
python-multipart==0.0.20
email-validator==2.2.0
resend==2.4.0
openai==1.59.6
